Fix rebuild by closing stream
Problem: Previously I thought that maybe we shouldn't be setting `since`
to `-1`, but that's not the real problem. The root of the issue is that
we aren't closing the stream, which means that the stream doesn't get
started by FlumeDB.

Solution: Undo previous `since` fix and correctly close the stream
during `destroy()`, plus add tests ensuring that the `rebuild()` method
is working correctly in this module.